Oops... Forgot to compile before previous commit, contained copy-paste bug.
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@6590 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/FixedPoint.hh
@@ -195,7 +195,7 @@ private:
 // Force all constants being defined, some compilers need this:
 
 template <unsigned FRACTION_BITS>
-const float FixedPoint<FRACTION_BITS>::ONE;
+const int FixedPoint<FRACTION_BITS>::ONE;
 
 template <unsigned FRACTION_BITS>
 const float FixedPoint<FRACTION_BITS>::INV_ONE_F;
